+++ src/app-pages/Reports/CorpOfficeReports/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import TextSection from '../../../app-components/TextSection';
-import ReportsContainer from "../../../app-components/Reports";
+import ReportsContainer from "../ReportsContainer";
 import { connect } from "redux-bundler-react";
 import { RoutePaths } from "../../../app-bundles/routes-bundle";
